:truck: [templates] Move extensions module from application.cookiecutter
